refactor: move document, labels, and parameters to elements/
- Move document.rs to elements/document.rs (Document element parser)
- Move labels.rs to elements/labels.rs (annotation label parsing)
- Move parameters.rs to elements/parameters.rs (annotation parameter parsing)
- Update all imports to use elements:: paths
- Update module declarations in parser.rs

All element-bound parsing code is now co-located in elements/ directory.
All 181 tests pass (178 passed, 3 ignored).
